refactor: 좋아요 상태가 true 인 게시글 좋아요 개수 반환하도록 변경
- 테스트 추가

+++ src/main/java/commaproject/be/commaserver/service/dto/CommaDetailResponse.java
@@ -13,11 +13,11 @@ public class CommaDetailResponse {
     private String username;
     private Long userId;
     private LocalDateTime createdAt;
-    private int postLikeCount;
+    private Long postLikeCount;
     private List<CommentDetailResponse> comments;
 
     public CommaDetailResponse(Long id, String title, String content, String username, Long userId,
-        LocalDateTime createdAt, int postLikeCount, List<CommentDetailResponse> comments) {
+        LocalDateTime createdAt, Long postLikeCount, List<CommentDetailResponse> comments) {
         this.id = id;
         this.title = title;
         this.content = content;
